Optimize pairwise merging in q_merge using two-pointer approach
Use a two-pointer strategy to merge queue pairs from the front and
back of the list. This reduces the total number of queues by half in
each iteration. After each pass, reset the front pointer to the head
for the next round of merging.

The time complexity reduce from O(nlogn) to O(n).

Change-Id: Id5c30be1601a3da303591cabe51175dfb14a55e9