Implement Fisher–Yates shuffle for queue elements
Applies the modern Fisher–Yates algorithm to randomly shuffle the
elements in the queue. Uses a value-swap approach to preserve list
structure while achieving randomness.

Notice that queue.h is prohibited to modify, so before the function
is push to the git server, the corresponding code for shuffle in
queue.h will be deleted, and the code for shuffle in qtest.c will be
marked as comment.

Change-Id: I5198da1e74bc9eceeec09408b3056bb8aeabe366